Add Chinese document.
